third_party/blink/web_tests/platform/generic/fast/block/block-parent-with-zero-width-child-expected.txt

This tests if the parent block has the correct size when the child element has fixed width as zero.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS successfullyParsed is true

TEST COMPLETE

PASS document.getElementById("p_topRight").getClientRects()[0].width  is 50
PASS document.getElementById("p_botRight").getClientRects()[0].width  is 50
PASS successfullyParsed is true

TEST COMPLETE
